STM32 fixes for F4 32-bit timers
git-svn-id: https://nuttx.svn.sourceforge.net/svnroot/nuttx/trunk@4300 7fd9a85b-ad96-42d3-883c-3090e2eb8679

+++ nuttx/arch/arm/src/stm32/stm32_tim.h
@@ -6,7 +6,7 @@
  *
  * With modifications and updates by:
  *
- *   Copyright (C) 2011 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2011-2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -92,7 +92,7 @@ struct stm32_tim_dev_s
 typedef enum
 {
   STM32_TIM_MODE_UNUSED       = -1,
-    
+
   /* One of the following */
 
   STM32_TIM_MODE_MASK         = 0x0310,
@@ -101,7 +101,7 @@ typedef enum
   STM32_TIM_MODE_DOWN         = 0x0110,
   STM32_TIM_MODE_UPDOWN       = 0x0200,
   STM32_TIM_MODE_PULSE        = 0x0300,
-    
+
   /* One of the following */
 
   STM32_TIM_MODE_CK_INT       = 0x0000,
@@ -116,9 +116,9 @@ typedef enum
 //STM32_TIM_MODE_CK_CH2       = 0x0002,
 //STM32_TIM_MODE_CK_CH3       = 0x0003,
 //STM32_TIM_MODE_CK_CH4       = 0x0004
-    
+
   /* Todo: external trigger block */
-    
+
 } stm32_tim_mode_t;
 
 /* TIM Channel Modes */
@@ -126,26 +126,26 @@ typedef enum
 typedef enum
 {
   STM32_TIM_CH_DISABLED       = 0x00,
-    
+
   /* Common configuration */
 
   STM32_TIM_CH_POLARITY_POS   = 0x00,
   STM32_TIM_CH_POLARITY_NEG   = 0x01,
-    
+
   /* MODES: */
 
   STM32_TIM_CH_MODE_MASK      = 0x06,
-    
+
   /* Output Compare Modes */
 
   STM32_TIM_CH_OUTPWM         = 0x04,     /** Enable standard PWM mode, active high when counter < compare */
 //STM32_TIM_CH_OUTCOMPARE     = 0x06,
-    
+
   // TODO other modes ... as PWM capture, ENCODER and Hall Sensor
 //STM32_TIM_CH_INCAPTURE      = 0x10,
 //STM32_TIM_CH_INPWM          = 0x20
 //STM32_TIM_CH_DRIVE_OC   -- open collector mode
-    
+
 } stm32_tim_channel_t;
 
 /* TIM Operations */
@@ -156,12 +156,12 @@ struct stm32_tim_ops_s
 
   int  (*setmode)(FAR struct stm32_tim_dev_s *dev, stm32_tim_mode_t mode);
   int  (*setclock)(FAR struct stm32_tim_dev_s *dev, uint32_t freq);
-  void (*setperiod)(FAR struct stm32_tim_dev_s *dev, uint16_t period);
-    
+  void (*setperiod)(FAR struct stm32_tim_dev_s *dev, uint32_t period);
+
   /* General and Advanced Timers Adds */
-    
+
   int  (*setchannel)(FAR struct stm32_tim_dev_s *dev, uint8_t channel, stm32_tim_channel_t mode);
-  int  (*setcompare)(FAR struct stm32_tim_dev_s *dev, uint8_t channel, uint16_t compare);
+  int  (*setcompare)(FAR struct stm32_tim_dev_s *dev, uint8_t channel, uint32_t compare);
   int  (*getcapture)(FAR struct stm32_tim_dev_s *dev, uint8_t channel);
 
   int  (*setisr)(FAR struct stm32_tim_dev_s *dev, int (*handler)(int irq, void *context), int source);
